+++ test/XmlTools.Tests/CodeGenerator/DecimalAndIntegerTypeCorrectorTests.cs
@@ -1,10 +1,11 @@
-﻿using System.Xml.Linq;
+﻿using System.Linq;
+using System.Xml.Linq;
 using XmlTools.Tests.CodeGenerator.FileTests;
 using Xunit;
 
 namespace XmlTools.Tests.CodeGenerator
 {
-    public class DecimalTypeCorrectorTests
+    public class DecimalAndIntegerTypeCorrectorTests
     {
         [Theory]
         [InlineData("1", "1")]
@@ -52,6 +53,127 @@ namespace XmlTools.Tests.CodeGenerator
             var expectedXDoc = XDocument.Parse(GetXmlStringWithValue(expectedValue));
             var xDocComparator = new XDocumentComparator(expectedXDoc, correctedXDoc);
             xDocComparator.AssertXDocumentsAreEqual();
+		}
+
+		[Theory]
+        [InlineData("Hello World!", false, "")]
+        [InlineData("d", false, "")]
+        [InlineData("1", true, "")]
+        [InlineData("1.1", true, "")]
+        [InlineData("999.999", true, "")]
+        [InlineData("999,999.999", true, "999999.999")]
+        [InlineData("1,23", true, "")]
+        public void RemovesInvalidValuesInDecimalAttributes(string replacement, bool shouldKeepElement, string expectedValue)
+        {
+            var input = @"<?xml version=""1.0"" encoding=""UTF-8""?>
+<GAEB xmlns=""http://www.gaeb.de/GAEB_DA_XML/200407"">
+    <GAEBInfo>
+        <Version>3.1</Version>
+        <VersDate>2009-12</VersDate>
+    </GAEBInfo>
+    <Award>
+        <DP>83</DP>
+        <BoQ>
+            <BoQBody>
+                <BoQCtgy RNoPart=""1"">
+                    <BoQBody>
+                        <Itemlist>
+                            <Item RNoPart=""2"">
+                                <Qty>59</Qty>
+                                <QU>stck</QU>
+                                <Description>
+                                    <CompleteText>
+                                        <DetailTxt>
+                                            <TextComplement>
+                                                <ComplBodyDec Value=""@@replacement@@""/>
+                                            </TextComplement>
+                                        </DetailTxt>
+                                    </CompleteText>
+                                </Description>
+                            </Item>
+                        </Itemlist>
+                    </BoQBody>
+                </BoQCtgy>
+            </BoQBody>
+        </BoQ>
+    </Award>
+</GAEB>"
+.Replace("@@replacement@@", replacement);
+
+            var schemaFile = ParserTestFile.GAEB_XML_3_1_Schema;
+            var invalidXDoc = XDocument.Parse(input);
+            var correctedXDoc = SchemaCorrectorHelper.CorrectXmlInstanceForSchema(schemaFile, invalidXDoc);
+            var expectedXDoc = XDocument.Parse(input);
+			var expectedNode = expectedXDoc.Descendants().Single(d => d.Name.LocalName == "ComplBodyDec");
+            if (!shouldKeepElement)
+			{
+				expectedNode.Attribute("Value").Remove();
+            }
+			if (!string.IsNullOrWhiteSpace(expectedValue))
+            {
+				expectedNode.Attribute("Value").Value = expectedValue;
+            }
+            var xDocComparator = new XDocumentComparator(expectedXDoc, correctedXDoc);
+            xDocComparator.AssertXDocumentsAreEqual();
+		}
+
+		[Theory]
+		[InlineData("Hello World!", false)]
+		[InlineData("d", false)]
+		[InlineData("1", true)]
+		[InlineData("1.1", false)]
+		[InlineData("999.999", false)]
+		[InlineData("999,999.999", false)]
+		[InlineData("999999999", true)]
+		[InlineData("1,23", false)]
+		public void RemovesInvalidValuesInIntegerAttributes(string replacement, bool shouldKeepElement)
+		{
+			var input = @"<?xml version=""1.0"" encoding=""UTF-8""?>
+<GAEB xmlns=""http://www.gaeb.de/GAEB_DA_XML/200407"">
+    <GAEBInfo>
+        <Version>3.1</Version>
+        <VersDate>2009-12</VersDate>
+    </GAEBInfo>
+    <Award>
+        <DP>83</DP>
+        <BoQ>
+            <BoQBody>
+                <BoQCtgy RNoPart=""1"">
+                    <BoQBody>
+                        <Itemlist>
+                            <Item RNoPart=""2"">
+                                <Qty>59</Qty>
+                                <QU>stck</QU>
+                                <Description>
+                                    <CompleteText>
+                                        <DetailTxt>
+                                            <TextComplement>
+                                                <ComplBodyInt Value=""@@replacement@@""/>
+                                            </TextComplement>
+                                        </DetailTxt>
+                                    </CompleteText>
+                                </Description>
+                            </Item>
+                        </Itemlist>
+                    </BoQBody>
+                </BoQCtgy>
+            </BoQBody>
+        </BoQ>
+    </Award>
+</GAEB>"
+.Replace("@@replacement@@", replacement);
+
+			var schemaFile = ParserTestFile.GAEB_XML_3_1_Schema;
+			var invalidXDoc = XDocument.Parse(input);
+			var correctedXDoc = SchemaCorrectorHelper.CorrectXmlInstanceForSchema(schemaFile, invalidXDoc);
+			var expectedXDoc = XDocument.Parse(input);
+			var expectedNode = expectedXDoc.Descendants().Single(d => d.Name.LocalName == "ComplBodyInt");
+			if (!shouldKeepElement)
+			{
+				expectedNode.Attribute("Value").Remove();
+			}
+			var xDocComparator = new XDocumentComparator(expectedXDoc, correctedXDoc);
+			xDocComparator.AssertXDocumentsAreEqual();
 		}
 
 		[Theory]